requirements.txt
pillow
pdf2image
poppler-utils  # Required for pdf2image to work (Linux)
